Use additionalinputsmixin for job create form
