+++ app/grandchallenge/algorithms/forms.py
@@ -60,13 +60,11 @@ from grandchallenge.algorithms.serializers import (
     AlgorithmSerializer,
 )
 from grandchallenge.algorithms.tasks import import_remote_algorithm_image
-from grandchallenge.components.form_fields import (
-    INTERFACE_FORM_FIELD_PREFIX,
-    InterfaceFormFieldFactory,
+from grandchallenge.components.forms import (
+    AdditionalInputsMixin,
+    ContainerImageForm,
 )
-from grandchallenge.components.forms import ContainerImageForm
 from grandchallenge.components.models import (
-    CIVData,
     ComponentInterface,
     ComponentJob,
     ImportStatusChoices,
@@ -111,7 +109,7 @@ class ModelFactsTextField(Field):
     template = "algorithms/model_facts_field.html"
 
 
-class JobCreateForm(SaveFormInitMixin, Form):
+class JobCreateForm(SaveFormInitMixin, AdditionalInputsMixin, Form):
     algorithm_image = ModelChoiceField(
         queryset=None, disabled=True, required=True, widget=HiddenInput
     )
@@ -162,30 +160,7 @@ class JobCreateForm(SaveFormInitMixin, Form):
             )
             self.fields["algorithm_model"].initial = active_model
 
-        for algorithm_input in interface.inputs.all():
-            prefixed_interface_slug = (
-                f"{INTERFACE_FORM_FIELD_PREFIX}{algorithm_input.slug}"
-            )
-
-            if prefixed_interface_slug in self.data:
-                if (
-                    not algorithm_input.requires_file
-                    and algorithm_input.kind == ComponentInterface.Kind.ANY
-                ):
-                    # interfaces for which the data can be a list need
-                    # to be retrieved with getlist() from the QueryDict
-                    initial = self.data.getlist(prefixed_interface_slug)
-                else:
-                    initial = self.data[prefixed_interface_slug]
-            else:
-                initial = None
-
-            self.fields[prefixed_interface_slug] = InterfaceFormFieldFactory(
-                interface=algorithm_input,
-                user=self._user,
-                required=algorithm_input.value_required,
-                initial=initial if initial else algorithm_input.default_value,
-            )
+        self.init_additional_inputs(inputs=interface.inputs.all())
 
     @cached_property
     def jobs_limit(self):
@@ -203,7 +178,7 @@ class JobCreateForm(SaveFormInitMixin, Form):
         if self.jobs_limit < 1:
             raise ValidationError("You have run out of algorithm credits")
 
-        cleaned_data = self.reformat_inputs(cleaned_data=cleaned_data)
+        cleaned_data["inputs"] = self.clean_additional_inputs()
 
         if Job.objects.get_jobs_with_same_inputs(
             inputs=cleaned_data["inputs"],
@@ -214,26 +189,6 @@ class JobCreateForm(SaveFormInitMixin, Form):
                 "A result for these inputs with the current image "
                 "and model already exists."
             )
-
-        return cleaned_data
-
-    def reformat_inputs(self, *, cleaned_data):
-        keys_to_remove = []
-        inputs = []
-        for k, v in cleaned_data.items():
-            if k.startswith(INTERFACE_FORM_FIELD_PREFIX):
-                keys_to_remove.append(k)
-                inputs.append(
-                    CIVData(
-                        interface_slug=k[len(INTERFACE_FORM_FIELD_PREFIX) :],
-                        value=v,
-                    )
-                )
-
-        for key in keys_to_remove:
-            cleaned_data.pop(key)
-
-        cleaned_data["inputs"] = inputs
 
         return cleaned_data
